file-classifier/requirements.txt
# API Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# LLM
openai==1.58.1

# File Extraction
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.1.0
chardet==5.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
structlog==24.4.0
tiktoken==0.8.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
